fix: renamed imports, moved classes to directories
