--- core/src/main/java/org/auSpherical/auFight/Collidable.java
@@ -1,4 +1,0 @@
-package org.auSpherical.auFight;
-
-public interface Collidable {
-}
