--- core/src/main/java/org/auSpherical/auFight/CollitionBox.java
@@ -1,4 +1,0 @@
-package org.auSpherical.auFight;
-
-public class CollitionBox {
-}
