+++ core/src/main/java/org/auSpherical/auFight/GameLogicManager.java
@@ -1,6 +1,9 @@
 //esta clase sirve para hacer la logica de los sprites
 package org.auSpherical.auFight;
 
+import org.auSpherical.auFight.inputs.KeyboardInput;
+import org.auSpherical.auFight.inputs.Player;
+
 public class GameLogicManager {
 
     AuFight main;
