+++ core/src/main/java/org/auSpherical/auFight/placeholders/Collidable.java
@@ -1,0 +1,4 @@
+package org.auSpherical.auFight.placeholders;
+
+public interface Collidable {
+}
