+++ core/src/main/java/org/auSpherical/auFight/placeholders/CollitionBox.java
@@ -1,0 +1,4 @@
+package org.auSpherical.auFight.placeholders;
+
+public class CollitionBox {
+}
