+++ core/src/main/java/org/auSpherical/auFight/placeholders/Damageable.java
@@ -1,4 +1,4 @@
-package org.auSpherical.auFight;
+package org.auSpherical.auFight.placeholders;
 
 public interface Damageable {
     void receiveDamage(int damage);
