+++ core/src/main/java/org/auSpherical/auFight/placeholders/Hitbox.java
@@ -1,6 +1,6 @@
-package org.auSpherical.auFight;
+package org.auSpherical.auFight.placeholders;
 import com.badlogic.gdx.math.Rectangle;
-import com.badlogic.gdx.math.collision.BoundingBox;
+
 public class Hitbox {
     int damage;
     Rectangle hitbox;
