+++ core/src/main/java/org/auSpherical/auFight/placeholders/Weapon.java
@@ -1,4 +1,6 @@
-package org.auSpherical.auFight;
+package org.auSpherical.auFight.placeholders;
+
+import org.auSpherical.auFight.Entity;
 
 public class Weapon extends Entity {
     private int activeFrame;
